Day 1 commit
Created a basic window and a tile set that is all "floor" tiles by creating a 2D list of Tile objects. Tiles are drawn in a 30x40 grid. Used floor and wall images from previous project in my Tile object.

Also started user interaction. User can click on a tile, and it will tell you which tell was clicked via a print statement.