Day 2 commit
Added the special "start/end" tile image. Allows the user to right click on a tile, which will set it as a start/end tile.

Next, allow the user to only set 2 tiles as a start or end tile. Will likely need to create a "World" object for this to keep track of which tiles are for which purpose.