Day 3 commit
Re-organized functions and created a World object to handle all the logic around creating and accessing the world map that is being created. Implemented not allowing the user to create more than 2 start or end tile.

Also added a "Save" button on the right hand side of the screen along with the save button image.

Next up, figure out how to create a text field in Python so the user can type in a name for their world and save the world to a file.