Day 4 commit
Created a text input field for the file name to save the world. Currently, the saving portion doesn't work, but the text field is functional.

Created an empty rectangle to represent the text field. When the user clicks on the text field, it becomes "active" and highlighted. If the text field is highlighted, any key presses will be displayed in the text field. If backspace is pressed, it will remove the last letter in the file name.

Next up, saving the world layout to a file where the walls are "#", floor is "." and start / end tiles are "S" and "E" respectively.