5/13/2024 commit
Loading maps is functional now. Added a "Load" button. When this button is clicked, it will call the "load_world" function. This function reads the file which is passed as parameter, and updates the tile_type of each tile based on what is in the file.

Next up, allowing the user to change multiple tiles at once by click and dragging (I'm not sure how to do this).