Click/dragging update
Figuring out how to do the click and drag to change multiple tiles. This is what I have so far:

When the user clicks down, it records the point clicked. When the user releases the mouse, it records the point where it was released. If those points aren't the same, then the user dragged the mouse.

If the user dragged the mouse, it is determining whether the line is "mostly" horizontal or "mostly" vertical by looking at the difference between the x coordinates and y coordinates.

Next up, determine which direction the user dragged (left, right, up or down), and get a list of points.