5/17/24 commit
Added selectable Coin sprite that will be highlighted when clicked on. Goal is for this to be able to allow the user to place Coins on the map in addition to switching tiles. Spent a lot of time fixing a bug that changed the bottom right tile whenever you click anywhere outside the map.