5/21/24 Commit
User can now place coins on the map by clicking on the coin sprite and clicking on tiles where the coin should be placed. If a coin is clicked on while the sprite is active, it will toggle coins on and off on a specific tile.

The save and load features also work with keeping track of where coins were placed and will draw them on the map in their appropriate place after loading.