5/28/24 commit
Added a selectable Enemy sprite that behaves the same way as the Coin sprite. The user cannot place enemies on the map yet.

Re-organized the "Coin" class. Renamed the class to a TileItem that allows you specify whether the object represents a coin or an enemy. Also moved the boolean variable to tell if the tile item is active into the class.